backend/requirements.txt
Flask>=3.0.0
SQLAlchemy==1.4.23
psycopg2-binary==2.9.5
Flask-SQLAlchemy>=3.0.0
ijson==3.2.3
pytest==8.3.3
Flask-Security==5.5.2
python-dotenv==1.0.1
argon2-cffi==23.1.0
flask-cors==4.0.0
flask-wtf==1.2.1